requirements.txt
memvid
PyPDF2
openai
python-dotenv
fastapi
uvicorn
pydantic